+++ Assets/PotatoUtil/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2022-04-22
+### Added
+- `FNVString` that is a version of the `FNVHash` that maintains the string value and uses it for serialization
+- restored `FNVHash`'s ability to serialize itself as an `uint`
+
 ## [2.2.0] - 2022-04-20
 ### Added
 - `Database` data structure that stores data by `FNVHash`